v0.8.2 Added a fallback for people without a block logging plugin

+++ com/minebans/pluginInterfaces/minebansantispam/MineBansAntiSpamPluginInterface.java
@@ -32,8 +32,8 @@ public class MineBansAntiSpamPluginInterface extends AntiSpamPluginInterface {
 		return true;
 	}
 	
-	public long getMaxViolationLevel(String playerName){
-		return (this.detectedSpammers.contains(playerName.toLowerCase())) ? 20L : 0L;
+	public int getMaxViolationLevel(String playerName){
+		return (this.detectedSpammers.contains(playerName.toLowerCase())) ? 20 : 0;
 	}
 	
 }
